services/lib/kafka_util/requirements.txt
django
django-kafka-consumer

# local libs: `python setup.py develop` from project root
# scheme_registry
